# src/api/v0/api.yaml
openapi: "3.0.0"
info:
  description: "Genomes on a Tree OpenAPI test."
  version: "0.0.1"
  title: "GoaT"
  contact:
    name: "GoaT"
    email: "goat@genomehubs.org"
  license:
    name: "MIT License"
servers:
  - url: "http://localhost:3000/api/v0.0.1"
tags:
  - name: "GoaT API"
    description: "Query Genomes on a Tree"
paths:
  /lookup:
    get:
      tags:
        - "GoaT API"
      summary: "Lookup"
      description: "Returns suggestions/identifiers matching a searchTerm"
      operationId: "getIdentifiers"
      x-eov-operation-handler: "api/v0/routes/lookup"
      parameters:
        - $ref: "#/components/parameters/searchTermParam"
        - $ref: "#/components/parameters/resultParam"
        - $ref: "#/components/parameters/prefixQueryParam"
        - $ref: "#/components/parameters/sizeParam"
        - $ref: "#/components/parameters/lineageParam"
        - $ref: "#/components/parameters/taxonomyParam"
        - $ref: "#/components/parameters/hubParam"
        - $ref: "#/components/parameters/versionParam"
        - $ref: "#/components/parameters/suggestSizeParam"
        - $ref: "#/components/parameters/gramSizeParam"
        - $ref: "#/components/parameters/maxErrorsParam"
        - $ref: "#/components/parameters/confidenceParam"
        - $ref: "#/components/parameters/preTagParam"
        - $ref: "#/components/parameters/postTagParam"
        - $ref: "#/components/parameters/indentParam"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                anyOf:
                  - status:
                    $ref: "#/components/schemas/Status"
                  - results:
                    $ref: "#/components/schemas/Results"
                  - suggestions:
                    $ref: "#/components/schemas/Suggestions"
        "400":
          description: "Invalid queryTerm"
  /record:
    get:
      tags:
        - "GoaT API"
      summary: "Returns individual records"
      description: "Returns records matching a record identifier"
      operationId: "getRecords"
      x-eov-operation-handler: "api/v0/routes/record"
      parameters:
        - $ref: "#/components/parameters/recordIdParam"
        - $ref: "#/components/parameters/resultParam"
        - $ref: "#/components/parameters/sizeParam"
        - $ref: "#/components/parameters/taxonomyParam"
        - $ref: "#/components/parameters/hubParam"
        - $ref: "#/components/parameters/versionParam"
        - $ref: "#/components/parameters/indentParam"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                anyOf:
                  - status:
                    $ref: "#/components/schemas/Status"
                  - results:
                    $ref: "#/components/schemas/Records"
        "400":
          description: "Invalid recordId"
  /search:
    get:
      tags:
        - "GoaT API"
      summary: "Search for records"
      description: "Returns records matching a query"
      operationId: "getSearchResults"
      x-eov-operation-handler: "api/v0/routes/search"
      parameters:
        - $ref: "#/components/parameters/queryParam"
        - $ref: "#/components/parameters/fieldsParam"
        - $ref: "#/components/parameters/resultParam"
        - $ref: "#/components/parameters/sizeParam"
        - $ref: "#/components/parameters/taxonomyParam"
        - $ref: "#/components/parameters/hubParam"
        - $ref: "#/components/parameters/versionParam"
        - $ref: "#/components/parameters/indentParam"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                anyOf:
                  - status:
                    $ref: "#/components/schemas/Status"
                  - results:
                    $ref: "#/components/schemas/Records"
        "400":
          description: "Invalid query"
  /count:
    get:
      tags:
        - "GoaT API"
      summary: "Count records"
      description: "Returns count of records matching a query"
      operationId: "getSearchResultCount"
      x-eov-operation-handler: "api/v0/routes/count"
      parameters:
        - $ref: "#/components/parameters/queryParam"
        - $ref: "#/components/parameters/resultParam"
        - $ref: "#/components/parameters/taxonomyParam"
        - $ref: "#/components/parameters/hubParam"
        - $ref: "#/components/parameters/versionParam"
        - $ref: "#/components/parameters/indentParam"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                anyOf:
                  - status:
                    $ref: "#/components/schemas/Status"
        "400":
          description: "Invalid query"
  /summary:
    get:
      tags:
        - "GoaT API"
      summary: "Search result summary"
      description: "Returns summary of search results"
      operationId: "getSearchSummary"
      x-eov-operation-handler: "api/v0/routes/summary"
      parameters:
        - $ref: "#/components/parameters/queryParam"
        - $ref: "#/components/parameters/summaryParam"
        - $ref: "#/components/parameters/resultParam"
        - $ref: "#/components/parameters/taxonomyParam"
        - $ref: "#/components/parameters/hubParam"
        - $ref: "#/components/parameters/versionParam"
        - $ref: "#/components/parameters/indentParam"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                anyOf:
                  - status:
                    $ref: "#/components/schemas/Status"
                  - results:
                    $ref: "#/components/schemas/Summaries"
        "400":
          description: "Invalid query"
components:
  parameters:
    searchTermParam:
      name: searchTerm
      in: query
      schema:
        type: string
      required: true
      description: Search term
    recordIdParam:
      name: recordId
      in: query
      schema:
        type: array
        items:
          type: string
      style: form
      explode: false
      required: true
      description: Record identifier
    queryParam:
      name: query
      in: query
      schema:
        type: string
      required: true
      description: Search query
    fieldsParam:
      name: fields
      in: query
      schema:
        type: string
        default: all
      required: false
      description: Comma separated list of fields to return
    resultParam:
      name: result
      in: query
      schema:
        type: string
        enum:
          - taxon
      required: true
      description: The result type (data set) to search against
    taxonomyParam:
      name: taxonomy
      in: query
      schema:
        type: string
        enum:
          - ncbi
      required: true
      description: The taxonomy to search against
    hubParam:
      name: hub
      in: query
      schema:
        type: string
        enum:
          - demo
      required: true
      description: GenomeHub name
    versionParam:
      name: version
      in: query
      schema:
        type: string
        enum:
          - "v0.1"
      required: true
      description: GenomeHub version number
    summaryParam:
      name: summary
      in: query
      schema:
        type: array
        items:
          type: string
          enum:
            - histogram
      required: true
      description: The summary type to return
    prefixQueryParam:
      name: prefixQuery
      in: query
      schema:
        type: boolean
      description: Treat search term as a prefix
    lineageParam:
      name: lineage
      in: query
      schema:
        type: string
      required: false
      description: Lineage (name or taxon ID) to restrict search to
    suggestParam:
      name: suggest
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      required: false
      description: Number of valid terms to suggest
      style: form
      explode: false
    indentParam:
      name: indent
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
        maximum: 8
      required: false
      description: Format output with indented lines
      style: form
      explode: false
    sizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 10
        minimum: 0
        maximum: 1000
      required: false
      description: Maximum number of results to return
      style: form
      explode: false
    suggestSizeParam:
      name: suggestSize
      in: query
      schema:
        type: integer
        default: 3
        minimum: 0
        maximum: 10
      required: false
      description: Number of similar terms to suggest
      style: form
      explode: false
    gramSizeParam:
      name: gramSize
      in: query
      schema:
        type: integer
        default: 3
        minimum: 2
        maximum: 4
      required: false
      description: Size of N-grams used by tokeniser
      style: form
      explode: false
    maxErrorsParam:
      name: maxErrors
      in: query
      schema:
        type: integer
        default: 3
        minimum: 0
        maximum: 10
      required: false
      description: Maximumum error percentage when matching terms
      style: form
      explode: false
    confidenceParam:
      name: confidence
      in: query
      schema:
        type: number
        default: 1
        minimum: 0
        maximum: 2
      required: false
      description: Threshold for candidate suggestions relative to input phrase
      style: form
      explode: false
    preTagParam:
      name: preTag
      in: query
      schema:
        type: string
        default: <em>
      required: false
      description: Tag to place before matching portion of query term
      style: form
      explode: false
    postTagParam:
      name: postTag
      in: query
      schema:
        type: string
        default: </em>
      required: false
      description: Tag to place after matching portion of query term
      style: form
      explode: false
  schemas:
    Status:
      type: object
      properties:
        success:
          type: boolean
        took:
          type: integer
          description: Time (ms) taken to execute query
        error:
          type: string
        hits:
          type: integer
    Records:
      type: array
      items:
        type: object
        properties:
          id:
            type: string
          index:
            type: string
          found:
            type: boolean
          record:
            type: object
    Results:
      type: array
      items:
        type: object
        properties:
          score:
            type: number
          result:
            type: object
          reason:
            type: array
            items:
              type: object
              properties:
                score:
                  type: number
                fields:
                  type: object
    Summaries:
      type: array
      items:
        type: object
        properties:
          name:
            type: string
          summary:
            type: object
    Suggestions:
      type: array
      items:
        type: object
        properties:
          text:
            type: string
          suggestion:
            type: object
            properties:
              text:
                type: string
              highlighted:
                type: string
              score:
                type: number
              collate_match:
                type: boolean